common: call is not common, structs are
