+++ crates/cli/Cargo.toml
@@ -14,4 +14,5 @@ argon2 = { version = "0.5.3", features = ["alloc","password-hash"] }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0.192", features = ["derive"] }
 chrono = { version="0.4.40", features = ["serde"] }
+reqwest = { version = "0.12", features = ["json"] }
 common = { workspace = true }
