+++ crates/cli/src/handlers/user/login.rs
@@ -1,9 +1,10 @@
+use common::structs::struct_user::{LoginAnswer, LoginPost};
+
 use crate::{
     API_URL, TOKEN,
+    call::{Answer, call},
     configuration::loader::{load, write_config},
-    structs::struct_user::{LoginAnswer, LoginPost},
 };
-use common::{Answer, call};
 
 impl Answer for LoginAnswer {
     fn code(&self) -> i32 {
