+++ crates/common/src/structs/struct_event.rs
